feat: SidebarContainer 컴포넌트 구현
